pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cloud_sre_v2"
version = "0.1.0"
description = "Cascading Incident Response RL Environment for OpenEnv — real microservices, real cascading failures"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # HTTP services
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.27.0",
    # Auth
    "pyjwt>=2.8.0",
    # Process monitoring
    "psutil>=5.9.0",
    # Metrics formatting
    "tabulate>=0.9.0",
    # Inference script
    "openai>=1.0.0",
]

[project.optional-dependencies]
train = [
    "trl[vllm]==0.29.0",
    "vllm==0.11.2",
    "torch>=2.8.0,<2.10.0",
    "peft",
    "transformers",
    "datasets",
    "matplotlib",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "cloud_sre_v2.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cloud_sre_v2", "cloud_sre_v2.server", "cloud_sre_v2.services", "cloud_sre_v2.infra"]
package-dir = { "cloud_sre_v2" = ".", "cloud_sre_v2.server" = "server", "cloud_sre_v2.services" = "services", "cloud_sre_v2.infra" = "infra" }
